requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-login==0.6.2
python-dotenv==1.0.0
openai-whisper==20231117
google-generativeai==0.3.2
gunicorn==21.2.0